platform: Passing and returning CHERI capability for input for fw_platform_init()
Change fw_platform_init() function type definitions to accpet CHERI
capability as input arguments and return for Zcheripurecap ABI.

+++ platform/generic/platform.c
@@ -139,9 +139,9 @@ default_config:
  * FDT is unchanged (or FDT is modified in-place) then fw_platform_init()
  * can always return the original FDT location (i.e. 'arg1') unmodified.
  */
-unsigned long fw_platform_init(unsigned long arg0, unsigned long arg1,
-				unsigned long arg2, unsigned long arg3,
-				unsigned long arg4)
+uintptr_t fw_platform_init(uintptr_t arg0, uintptr_t arg1,
+				uintptr_t arg2, uintptr_t arg3,
+				uintptr_t arg4)
 {
 	const char *model;
 	const void *fdt = (void *)arg1;
